gamefi_analysis/requirements.txt
requests
pandas
matplotlib
seaborn
ollama-python

